refactor: move and delete
Move code from peaks_in_matplotlib.py to trees.py.
Move code from hierarchical_peaks.py to utilities.py and delete hierarchical_peaks.py.
